requirements.txt
apache-airflow==2.7.3
pandas==2.1.0
requests==2.31.0
SQLAlchemy==2.0.21
